Skip side

+++ src/telomore/utils/classes_and_small_func.py
@@ -227,6 +227,8 @@ class Replicon:
             self.trim_map_index,
         ]
         for path in tmp_files:
+            if not path:
+                continue
             if os.path.exists(path):
                 os.remove(path)
 
@@ -283,7 +285,8 @@ class Replicon:
         if mode == 'nanopore':
             if os.path.exists(self.cons_log_np_out):
                 shutil.move(
-                    src=self.cons_log_np_out, dst=os.path.join(folder, self.cons_log_np_out)
+                    src=self.cons_log_np_out,
+                    dst=os.path.join(folder, self.cons_log_np_out),
                 )
         elif mode == 'illumina':
             if os.path.exists(self.cons_log_ill_out):
